+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - `msgtrace` module with ability to provide info such as date of last check for
-  updates or updater execution messages.
+  updates or updater execution messages
+- `packages` module that allows read access of installed packages info
 - Option `--no-network-fail` to skip updater run when network connection is
   unavailable
 - Possibility to disable network test by setting zero